// src/types/lo-report.ts
export interface LOReportData {
  id: string;
  outbound_delivery_number: string;
  outbound_delivery_item_number: string;
  lti: string;
  driver_name: string;
  vehicle_plate: string;
  transporter_name: string;
  material_description: string;
  storage_location_name: string;
  gate_number: string;
  batch_number: string;
  loading_date: string;
  created_at: string;
  destination: string;
  mt_net: number;
}